Task 1 :Iris Flowers Classification ML
1)
Project :
This particular ML project is usually
referred to as the "Hello World" of Machine
Learning. The iris flowers dataset contains
numeric attributes, and it is perfect for
beginners to learn about supervised ML
algorithms, mainly how to load and handle
data. Also, since this is a small dataset, it
can easily fit in memory without requiring
special
transformations or
scaling
capabilities.